perf: optimize QToken storage layout, structs and function arguments

+++ contracts/interfaces/IOracleRegistry.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 /// @author Rolla
 /// @notice oracle provider registry for holding a list of oracle providers and their id
 interface IOracleRegistry {
-    event AddedOracle(address oracle, uint256 oracleId);
+    event AddedOracle(address oracle, uint248 oracleId);
 
     event ActivatedOracle(address oracle);
 
@@ -14,7 +14,7 @@ interface IOracleRegistry {
     /// @notice Add an oracle to the oracle registry which will generate an id. By default oracles are deactivated
     /// @param _oracle the address of the oracle
     /// @return the id of the oracle
-    function addOracle(address _oracle) external returns (uint256);
+    function addOracle(address _oracle) external returns (uint248);
 
     /// @notice Deactivate an oracle so no new options can be created with this oracle address.
     /// @param _oracle the oracle to deactivate
@@ -25,7 +25,7 @@ interface IOracleRegistry {
     function activateOracle(address _oracle) external returns (bool);
 
     /// @notice oracle address => OracleInfo
-    function oracleInfo(address) external view returns (bool, uint256);
+    function oracleInfo(address) external view returns (bool, uint248);
 
     /// @notice exhaustive list of oracles in map
     function oracles(uint256) external view returns (address);
@@ -40,9 +40,9 @@ interface IOracleRegistry {
 
     /// @notice Get the numeric id of an oracle
     /// @param _oracle the oracle to get the id of
-    function getOracleId(address _oracle) external view returns (uint256);
+    function getOracleId(address _oracle) external view returns (uint248);
 
     /// @notice Get total number of oracles in registry
     /// @return the number of oracles in the registry
-    function getOraclesLength() external view returns (uint256);
+    function getOraclesLength() external view returns (uint248);
 }
